requirements.txt
pytest==8.3.5
Appium-Python-Client==4.5.1
opencv-python==4.11.0.86
selenium==4.29.0